requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
gTTS==2.2.2
idna==2.10
praw==7.2.0
prawcore==2.0.0
requests==2.25.1
six==1.15.0
update-checker==0.18.0
urllib3==1.26.4
websocket-client==0.58.0
Pillow==8.2.0
